upgrade libwebp to version 1.3.2
