+++ 3rdparty/libwebp/src/dec/vp8li_dec.h
@@ -126,19 +126,6 @@ void VP8LClear(VP8LDecoder* const dec);
 // Clears and deallocate a lossless decoder instance.
 void VP8LDelete(VP8LDecoder* const dec);
 
-// Helper function for reading the different Huffman codes and storing them in
-// 'huffman_tables' and 'htree_groups'.
-// If mapping is NULL 'num_htree_groups_max' must equal 'num_htree_groups'.
-// If it is not NULL, it maps 'num_htree_groups_max' indices to the
-// 'num_htree_groups' groups. If 'num_htree_groups_max' > 'num_htree_groups',
-// some of those indices map to -1. This is used for non-balanced codes to
-// limit memory usage.
-int ReadHuffmanCodesHelper(int color_cache_bits, int num_htree_groups,
-                           int num_htree_groups_max, const int* const mapping,
-                           VP8LDecoder* const dec,
-                           HuffmanTables* const huffman_tables,
-                           HTreeGroup** const htree_groups);
-
 //------------------------------------------------------------------------------
 
 #ifdef __cplusplus
