+++ 3rdparty/libwebp/src/enc/picture_enc.c
@@ -12,10 +12,10 @@
 // Author: Skal (pascal.massimino@gmail.com)
 
 #include <assert.h>
-#include <limits.h>
 #include <stdlib.h>
 
 #include "src/enc/vp8i_enc.h"
+#include "src/dsp/dsp.h"
 #include "src/utils/utils.h"
 
 //------------------------------------------------------------------------------
