+++ 3rdparty/libwebp/src/libwebp.pc.in
@@ -1,0 +1,12 @@
+prefix=@prefix@
+exec_prefix=@exec_prefix@
+libdir=@libdir@
+includedir=@includedir@
+
+Name: libwebp
+Description: Library for the WebP graphics format
+Version: @PACKAGE_VERSION@
+Requires.private: libsharpyuv
+Cflags: -I${includedir}
+Libs: -L${libdir} -l@webp_libname_prefix@webp
+Libs.private: -lm @PTHREAD_CFLAGS@ @PTHREAD_LIBS@
