+++ 3rdparty/libwebp/src/mux/muxedit.c
@@ -555,8 +555,7 @@ static WebPMuxError MuxCleanup(WebPMux* const mux) {
   if (num_frames == 1) {
     WebPMuxImage* frame = NULL;
     err = MuxImageGetNth((const WebPMuxImage**)&mux->images_, 1, &frame);
-    if (err != WEBP_MUX_OK) return err;
-    // We know that one frame does exist.
+    assert(err == WEBP_MUX_OK);  // We know that one frame does exist.
     assert(frame != NULL);
     if (frame->header_ != NULL &&
         ((mux->canvas_width_ == 0 && mux->canvas_height_ == 0) ||
